+++ Documentation/Changelog.md
@@ -2,6 +2,14 @@
 
 For all changes, please update changelog and use Year-Month-Day
 
+
+## 0.1.1
+2024.10.09
+- add pval threshold option
+- add pathway categories as an optional argument in gsea script
+- general documentation update
+- fix general wrapper `...` argument passing to just gsea script
+
 ## 0.1
 2024.10.09
 - initial commit, get gsea pipeline running reproducibly across runs